begin uplift of console project for DI
