+++ src/Common/Common.csproj
@@ -11,6 +11,7 @@
     <PackageReference Include="OpenTelemetry.Instrumentation.Http" Version="1.0.0-rc7" />
     <PackageReference Include="prometheus-net" Version="4.2.0" />
     <PackageReference Include="prometheus-net.AspNetCore" Version="4.1.1" />
+    <PackageReference Include="prometheus-net.DotNetRuntime" Version="4.1.0" />
     <PackageReference Include="Serilog" Version="2.10.0" />
     <PackageReference Include="System.Text.Json" Version="6.0.0-preview.3.21201.4" />
   </ItemGroup>
