+++ src/Common/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace Common
+{
+    public static class Constants
+    {
+        public const string AppName = "P2G";
+    }
+}
