--- src/Common/Logging.cs
@@ -1,9 +1,0 @@
-﻿using Serilog;
-
-namespace Common
-{
-	public class LogContext 
-	{
-		public static ILogger ForClass<T>() => Log.Logger.ForContext(Serilog.Core.Constants.SourceContextPropertyName, typeof(T).Name);
-	}
-}
