+++ src/Common/Observe/Logging.cs
@@ -1,0 +1,9 @@
+﻿using Serilog;
+
+namespace Common.Observe
+{
+    public class LogContext
+    {
+        public static ILogger ForClass<T>() => Log.Logger.ForContext(Serilog.Core.Constants.SourceContextPropertyName, typeof(T).Name);
+    }
+}
