+++ src/Garmin/GarminExceptions.cs
@@ -1,10 +1,9 @@
-﻿using Common;
-using Serilog;
+﻿using Serilog;
 using System;
 
 namespace Garmin
 {
-	public class GarminUploadException : Exception
+    public class GarminUploadException : Exception
 	{
 		private static readonly ILogger _logger = LogContext.ForClass<GarminUploadException>();
