+++ src/Garmin/GarminUploader.cs
@@ -12,7 +12,7 @@ using Metrics = Prometheus.Metrics;
 
 namespace Garmin
 {
-	public interface IGarminUploader
+    public interface IGarminUploader
 	{
 		Task UploadToGarminAsync();
 	}
@@ -21,7 +21,7 @@ namespace Garmin
 	{
 		private static readonly Histogram WorkoutUploadDuration = Metrics.CreateHistogram("p2g_workout_upload_duration_seconds", "Histogram of workout upload durations.", new HistogramConfiguration()
 		{
-			LabelNames = new[] { Common.Metrics.Label.Count }
+			LabelNames = new[] { Common.Observability.Metrics.Label.Count }
 		});
 		private static readonly Gauge FailedUploadAttemptsGauge = Metrics.CreateGauge("p2g_failed_upload_attempts",
 			"The number of consecutive failed upload attempts. Resets to 0 on the first successful upload. This is not a count of the number of workouts that failed to upload. P2G uploads in bulk, so this is just a guage of number of failed upload attempts.");
