+++ src/WebApp/Services/BackgroundSyncJob.cs
@@ -13,17 +13,17 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
-using static Common.Metrics;
+using static Common.Observability.Metrics;
 using Metrics = Prometheus.Metrics;
 
 namespace WebApp.Services
 {
-	public class BackgroundSyncJob : BackgroundService
+    public class BackgroundSyncJob : BackgroundService
 	{
 		private static readonly Histogram SyncHistogram = Metrics.CreateHistogram("p2g_sync_duration_seconds", "The histogram of sync jobs that have run.");
 		private static readonly Gauge BuildInfo = Metrics.CreateGauge("p2g_build_info", "Build info for the running instance.", new GaugeConfiguration()
 		{
-			LabelNames = new[] { Common.Metrics.Label.Version, Common.Metrics.Label.Os, Common.Metrics.Label.OsVersion, Common.Metrics.Label.DotNetRuntime }
+			LabelNames = new[] { Common.Observability.Metrics.Label.Version, Common.Observability.Metrics.Label.Os, Common.Observability.Metrics.Label.OsVersion, Common.Observability.Metrics.Label.DotNetRuntime }
 		});
 		private static readonly Gauge Health = Metrics.CreateGauge("p2g_sync_service_health", "Health status for P2G Sync Service.");
 		private static readonly Gauge NextSyncTime = Metrics.CreateGauge("p2g_next_sync_time", "The next time the sync will run in seconds since epoch.");
